Start Implementing Design Patterns for Lab Four
I began with implementing the Observer Pattern. This involved using an Observer interface, so my changes could be kept continuous across my different panels including stats, chart, and table. When my chart panel used to update, it would go on a new frame, without disposing of the previous one. Now, I must fix the statsPanel, which is not working.